Combined feature files into their own directories and combined features files where tests were in the same area. Addressed agreed standards and move constant.py file to new location. Updated READ.ME. Set every tag to be @step and not @whens, etc.
Removing prefix THAT from start of sentances

Moved test_steps.py tests into correct file locations and removed

Refactoring or removing comments

Placing all tests into their own directories.

Combin feature files that can be combined or share the same testing areas, as well as new file structure.

Updated README.md with file structure and location

correcting import path for constansts.py

Combining cla_eligible and cla_not_eligible and cla_diagnosis and cla_assignments

File renames

Moving enviroment.py into correct directory to fix failing tests

combining python files

combining Python files to make feature files
