+++ behave/features/cla_backend/backend_reports.feature
@@ -1,9 +1,9 @@
 Feature: User journeys for Fox admin or cla_backend
 
 Background: Start page
-    Given that I am logged in as "CHS_GENERAL_USER"
-    And that I have created cases with callbacks
-    Then that I am logged in as "FOX_ADMIN_GENERAL_USER"
+    Given I am logged in as "CHS_GENERAL_USER"
+    And I have created cases with callbacks
+    Then I am logged in as "FOX_ADMIN_GENERAL_USER"
 
 #Journey P10  Navigate to reports page and download MI callback report (LGA-1860, LGA-1861)
 @fox-csv
