+++ behave/features/cla_frontend/frontend_existing_callback.feature
@@ -1,10 +1,19 @@
-Feature: Select Callbacks for Digital Cases
+Feature: Check Callbacks for Digital Cases
 # Looking at the callbacks calendar view to see whether callbacks have been created and then viewing the cases
 
 #As a CHS Operator, I want to select a callback slot, so I can see the cases where a callback is booked for that slot.
 Background: Create Callbacks
-    Given that I am logged in as "CHS_GENERAL_USER"
-    And that I have created cases with callbacks
+    Given I am logged in as "CHS_GENERAL_USER"
+    And I have created cases with callbacks
+
+
+@check_callback_exists
+Scenario: Can see callbacks booked in different slots
+    Given that I am on cases callback page
+    And multiple cases with a callback exists
+    When I select a callback slot
+    Then I can see the cases where a callback is booked for that slot
+
 
 @view_existing_callback
 Scenario: Can see details of selected callback
@@ -13,13 +22,10 @@ Scenario: Can see details of selected callback
     When I select a case created on CLA Public from the callback slot
     Then I am taken to the "call centre" case details page
     And I can view the client details of a case created on CLA Public
-    | details     |
-    | Full name   |
-    | Telephone   |
+        | details     |
+        | Full name   |
+        | Telephone   |
     When I select "Start Call"
     Then the call has started
     And I remove the callback
     Then this case is removed from callback list (calendar view)
-
-
-
