+++ behave/features/cla_public/public_eligible.feature
@@ -1,8 +1,10 @@
-Feature: In scope means tests and some with callback or confirmation details after successful means test
+Feature:
+  - In scope means tests and some with callback or confirmation details after successful means test
+  - Out of scope means tests
 
 Background: Start page
     Given I have selected the start now button on the start page
-    
+
 @cla-in-scope-callback
 Scenario: Complete callback form asking for callback
   Given I have passed the means test
@@ -55,7 +57,7 @@ Scenario: Complete in scope housing check
      | Are you or your partner (if you have one) aged 60 or over? | No     |
      | Do you have any savings or investments?                    | No     |
      | Do you have any valuable items worth over £500 each?       | No     |
-  And I click Continue
+  And I click continue
   And I am taken to the "Your property" page located on "/property"
   And I <answer> the <question>
      | question                                                   | answer |
@@ -66,7 +68,19 @@ Scenario: Complete in scope housing check
      | How much was your monthly mortgage repayment last month?   | 0      |
      | Do you rent out any part of this property?                 | No     |
      | Is your share of the property in dispute?                  | No     |
-  And I click Continue
+  And I click continue
   And I am taken to the "Review your answers" page located on "/review"
   And I click Confirm
   And I am taken to the "You’re unlikely to get legal aid" page located on "/result/refer/housing"
+
+# Out of scope means tests
+@cla-out-of-scope-family-divorce
+Scenario: Complete out of scope means test for family and divorce
+  Given I am taken to the "Choose the area you most need help with" page located on "/scope/diagnosis/"
+  When I select the category <category>
+    | category                                                        |
+    | Family                                                          |
+    | Divorce, separation or dissolution                              |
+    | Any other problem to do with divorce, separation or dissolution |
+  Then I am taken to the "Legal aid doesn’t cover all types of problem" page located on "/scope/refer/family"
+
