+++ behave/features/steps/frontend_existing_callback.py
@@ -1,10 +1,99 @@
+from cla_common.call_centre_availability import OpeningHours
+from cla_common.constants import OPERATOR_HOURS
+from helper.constants import CLA_CALLBACK_CASES, CLA_FRONTEND_URL
+from features.steps.cla_in_scope import assert_header_on_page
 from features.steps.common_steps import compare_client_details_with_backend
-from behave import then, when, given, step
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.common.exceptions import StaleElementReferenceException
 
+from behave import step
 
-@given("I am viewing a callback slot")
+import pytz
+
+
+def get_next_available_callback_slots():
+    # Need to get available slots from cla_common.
+    # Use OpeningHours class with correct setup so that it knows about bank holidays
+    operator_hours = OpeningHours(**OPERATOR_HOURS)
+    available_days_from_common = operator_hours.available_days(2)
+    all_available_slots = []
+    for day in available_days_from_common:
+        available_slots = operator_hours.time_slots(day.date())
+        all_available_slots.extend(available_slots)
+    # do every other slot as on the display they lump two slots together on website
+    slots_chosen = all_available_slots[:8:2]
+    slots_chosen.append(all_available_slots[0])
+    return slots_chosen
+
+
+@step("I have created cases with callbacks")
+# This is the background step that takes test cases and assigns a callback slot some time over the next few days
+def step_impl_created_cases_and_callbacks(context):
+    slots_chosen = get_next_available_callback_slots()
+    for index, case in enumerate(CLA_CALLBACK_CASES):
+        # don't create a callback for the case if there is already one for this case
+        callback_check = context.helperfunc.get_case_callback_details_from_backend(case)
+        # look for CB code in logs, if no logs or if  no 'code' about callbacks in logs then no callback created
+        callback_already_created = next(
+            (item for item in callback_check if item["code"] in ["CB1", "CB2", "CB3"]),
+            None,
+        )
+        if callback_already_created is None:
+            # slots passed to backend must be utc but times from cla_common are not timezone aware.
+            next_slot = pytz.timezone("Europe/London").localize(slots_chosen[index])
+            next_slot_utc = next_slot.astimezone(pytz.utc)
+            time_slot_start = next_slot_utc.strftime("%d/%m/%Y %H:%M")
+            case_reference = case
+            callback_json = {
+                "datetime": time_slot_start,
+                "notes": "",
+                "priority_callback": False,
+            }
+            # we don't try and add another callback after one created and
+            # Opening Hours returns only available slots
+            new_callback = context.helperfunc.update_case_callback_details(
+                case_reference, callback_json
+            )
+            message = (
+                f'Error for case {new_callback["case_reference"]}, '
+                f'data {new_callback["call_back_json"]} returned {new_callback["response_json"]}'
+            )
+            # return error if not 204 response
+            assert new_callback["response_status_code"] == 204, message
+
+
+@step("that I am on cases callback page")
+def step_impl_cases_callback_page(context):
+    start_page_url = f"{CLA_FRONTEND_URL}/call_centre/callbacks/"
+    context.helperfunc.open(start_page_url)
+    assert_header_on_page("Cases", context)
+
+
+@step("multiple cases with a callback exists")
+def step_impl_multiple_cases(context):
+    #  there will be at least one <a> element that shows callbacks are booked
+    callbacks_exist = context.helperfunc.find_many_by_class("CallbackMatrix-slot")
+    assert (
+        len(callbacks_exist) > 1
+    ), f"Found only {len(callbacks_exist)} callbacks, expected to have more than one."
+
+
+@step("I select a callback slot")
+def step_impl_select_callback_slot(context):
+    # look for the callbacks and click on the first one
+    callbacks = context.helperfunc.find_many_by_class("CallbackMatrix-slot")
+    callbacks[0].click()
+
+
+@step("I can see the cases where a callback is booked for that slot")
+def step_impl_callback_booked(context):
+    # look for the cases displayed alongside the calendar
+    list_cases = context.helperfunc.find_by_class("ListTable")
+    case_rows = list_cases.find_elements_by_xpath("//tbody/tr/td")
+    assert case_rows is not None
+
+
+@step("I am viewing a callback slot")
 def step_impl_view_callback(context):
     # this is actually a composite of several of the steps in another scenario
     context.execute_steps(
@@ -15,7 +104,7 @@ def step_impl_view_callback(context):
     )
 
 
-@given("callback slot contains a case created on CLA Public")
+@step("callback slot contains a case created on CLA Public")
 def step_impl_contains_case(context):
     xpath_string = (
         '//table[@class="ListTable"]/tbody/tr[./td/abbr[@title="WEB CASE"]]/td/a'
@@ -29,7 +118,7 @@ def step_impl_contains_case(context):
     context.selected_case_ref = cla_callback_case.text
 
 
-@when("I select a case created on CLA Public from the callback slot")
+@step("I select a case created on CLA Public from the callback slot")
 def step_impl_select_case(context):
     # find and click on the case in the callback list
     assert context.callback_case_element_link is not None, "No callback case to link to"
@@ -41,7 +130,6 @@ def step_on_case_details_page(context):
     case_id = context.selected_case_ref
     client_section = context.helperfunc.find_by_id("personal_details")
     compare_client_details_with_backend(context, case_id, client_section)
-    # and check that this shows the source is WEB
     element = "p"
     title_value = "Case source"
     xpath_string = f'//{element}[@title="{title_value}"]'
@@ -50,7 +138,7 @@ def step_on_case_details_page(context):
     ), "Case is not from WEB"
 
 
-@when('I select "Start Call"')
+@step('I select "Start Call"')
 def step_impl_start_call(context):
     callback_btn = context.helperfunc.find_by_css_selector("callback-status button")
     assert (
@@ -74,7 +162,7 @@ def step_impl_start_call(context):
     wait.until(wait_until_callback_is_started)
 
 
-@then("the call has started")
+@step("the call has started")
 def step_impl_call_has_started(context):
     assert context.selected_case_ref is not None, "Not viewing a case"
     case_logs = context.helperfunc.get_case_callback_details_from_backend(
@@ -97,7 +185,7 @@ def step_impl_call_has_started(context):
         assert note_element is not None, f"Could not find case log: {case_log['notes']}"
 
 
-@then("I remove the callback")
+@step("I remove the callback")
 def step_impl_remove_callback(context):
     callback_wrapper_element = context.helperfunc.find_by_css_selector(
         "callback-status"
@@ -129,7 +217,7 @@ def step_impl_remove_callback(context):
     wait.until(wait_until_callback_is_stopped)
 
 
-@then("this case is removed from callback list (calendar view)")
+@step("this case is removed from callback list (calendar view)")
 def step_impl_callback_removed(context):
     assert context.selected_case_ref is not None, "Case reference missing"
     callbacks = context.helperfunc.get_future_callbacks()
